Introduce lib/ directory, move STY files and rename
Collect 'implementation details' files to lib/, clearing up
the repository and bundling them nicely.
They should not be seen/touched by normal users of the template.

This loses the *.sty, aka LaTeX package functionalities, since
packages cannot be in subdirectories, but that is not a big
loss at all here.