Facilitate Docker usage in documentation
Add more examples, a clearer path and some clear
statements on why to use Docker.

I decided to 'deprecate' manual usage.
We now use minted, and as such have brought in
Python as a dependency.
This is just getting too much, and teaching about
Docker once vs. all the other manual steps should
be absolutely worth it in the long run.