Fix tables, introduce new package
One table broke because of siunitx version 3 and the
footnotes the table used.

Researching for replacements to tabu and its longtabu
table (which haven't been maintained in 10 years)
I found tabularray. A very promising, up and coming
package to do loads of stuff with tables. I implemented it
as an experiment.