Move class file to root
Solves a nasty issue: previously, we relied on 'acp.cls'
to be available in the image, at a tex-wide installation directory.
This had the advantage of that class being available to any document
using the image, but it had a core disadvantage: modifying the class file
would require *rebuilding the image*, which is big no bueno.
So even adjusting just the font weight on the title page, for example,
would necessitate an image rebuild.

Now, the class file is at the project root. Here, it can be found
by latex *directly*.
In order to still bake it into the image, the Docker build process
needs to be changed slightly (the build context needs to be moved up
to the root directory to find acp.cls, since the build context
cannot move up directories).
This gets us the best of both worlds with only a slightly more
annoying Docker build process.

Users can now change the class file freely, and changes will be
reflected in their document immediately, no image rebuild
necessary.